W-12204790: jdbc/jms driver specific cleanup not perfromed if driver is a sharedLibrary
Partial revert of
https://github.com/mulesoft/mule/commit/20aedfbc32f89867ecd59aa32ff605249c3d589a
plus test cases
