+++ modules/artifact/src/test/java/org/mule/runtime/module/artifact/classloader/ResourceReleaserTestCase.java
@@ -12,6 +12,7 @@ import static java.lang.Thread.currentThread;
 import static java.sql.DriverManager.deregisterDriver;
 import static java.sql.DriverManager.getDrivers;
 import static java.sql.DriverManager.registerDriver;
+import static java.util.Arrays.asList;
 import static java.util.Collections.list;
 import static java.util.Locale.getDefault;
 
@@ -28,6 +29,8 @@ import static org.mockito.Mockito.mock;
 import org.mule.module.artifact.classloader.JdbcResourceReleaser;
 import org.mule.runtime.module.artifact.api.classloader.ClassLoaderLookupPolicy;
 import org.mule.runtime.module.artifact.api.classloader.LookupStrategy;
+import org.mule.runtime.module.artifact.api.classloader.MuleArtifactClassLoader;
+import org.mule.runtime.module.artifact.api.classloader.MuleDeployableArtifactClassLoader;
 import org.mule.runtime.module.artifact.api.classloader.ResourceReleaser;
 import org.mule.runtime.module.artifact.api.descriptor.ArtifactDescriptor;
 import org.mule.runtime.module.artifact.internal.classloader.MulePluginClassLoader;
@@ -39,33 +42,57 @@ import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.net.URL;
 import java.sql.Driver;
+import java.util.Collection;
 import java.util.Map;
 import java.util.MissingResourceException;
 import java.util.ResourceBundle;
+import java.util.function.Function;
 import java.util.stream.Stream;
 
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.junit.runners.Parameterized.Parameters;
+
+import io.qameta.allure.Issue;
 
 @SmallTest
+@RunWith(Parameterized.class)
+@Issue("W-12204790")
 public class ResourceReleaserTestCase extends AbstractMuleTestCase {
 
   public static final String TEST_RESOURCE_RELEASER_CLASS_LOCATION =
       "/org/mule/runtime/module/artifact/classloader/TestResourceReleaser.class";
 
+  @Parameters
+  public static Collection<Function<ClassLoader, MuleArtifactClassLoader>> params() {
+    return asList(TestPluginClassLoader::new,
+                  // Libraries/drivers for which there is specific releaser code may be at the application/domain/policy level
+                  // rather than the plugin if they are added through `sharedLibraries` rather that
+                  // `additionalPluginDependencies`.
+                  TestApplicationClassLoader::new);
+  }
+
+  private final Function<ClassLoader, MuleArtifactClassLoader> classLoaderFactory;
+
+  public ResourceReleaserTestCase(Function<ClassLoader, MuleArtifactClassLoader> classLoaderFactory) {
+    this.classLoaderFactory = classLoaderFactory;
+  }
+
   @Test
   public void createdByCorrectArtifactClassLoader() throws Exception {
-    ensureResourceReleaserIsCreatedByCorrectClassLoader(new TestArtifactClassLoader(new TestArtifactClassLoader(currentThread()
+    ensureResourceReleaserIsCreatedByCorrectClassLoader(new TestPluginClassLoader(classLoaderFactory.apply(currentThread()
         .getContextClassLoader())));
   }
 
   @Test
   public void createdByCorrectParentArtifactClassLoader() throws Exception {
-    ensureResourceReleaserIsCreatedByCorrectClassLoader(new TestArtifactClassLoader(currentThread().getContextClassLoader()));
+    ensureResourceReleaserIsCreatedByCorrectClassLoader(classLoaderFactory.apply(currentThread().getContextClassLoader()));
   }
 
   @Test
   public void jdbcResourceReleaserShouldNotBeCreatedIfDriverIsNotLoaded() {
-    TestArtifactClassLoader testArtifactClassLoader = new TestArtifactClassLoader(currentThread().getContextClassLoader());
+    MuleArtifactClassLoader testArtifactClassLoader = classLoaderFactory.apply(currentThread().getContextClassLoader());
     testArtifactClassLoader.setResourceReleaserClassLocation(TEST_RESOURCE_RELEASER_CLASS_LOCATION);
     testArtifactClassLoader.dispose();
 
@@ -78,7 +105,7 @@ public class ResourceReleaserTestCase extends AbstractMuleTestCase {
 
   @Test
   public void jdbcResourceReleaserShouldNotBeCreatedIfDriverInterfaceIsLoaded() throws ClassNotFoundException {
-    TestArtifactClassLoader testArtifactClassLoader = new TestArtifactClassLoader(currentThread().getContextClassLoader());
+    MuleArtifactClassLoader testArtifactClassLoader = classLoaderFactory.apply(currentThread().getContextClassLoader());
     testArtifactClassLoader.setResourceReleaserClassLocation(TEST_RESOURCE_RELEASER_CLASS_LOCATION);
 
     testArtifactClassLoader.loadClass(Driver.class.getName());
@@ -97,8 +124,8 @@ public class ResourceReleaserTestCase extends AbstractMuleTestCase {
   @Test
   public void notDeregisterJdbcDriversDifferentClassLoaders() throws Exception {
     Driver jdbcDriver = mock(Driver.class);
-    TestArtifactClassLoader classLoader =
-        new TestArtifactClassLoader(new TestArtifactClassLoader(currentThread().getContextClassLoader()));
+    TestPluginClassLoader classLoader =
+        new TestPluginClassLoader(classLoaderFactory.apply(currentThread().getContextClassLoader()));
     try {
       registerDriver(jdbcDriver);
 
@@ -121,7 +148,7 @@ public class ResourceReleaserTestCase extends AbstractMuleTestCase {
     assertThat(list(getDrivers()), not(hasItem(jdbcDriver)));
   }
 
-  private void ensureResourceReleaserIsCreatedByCorrectClassLoader(MulePluginClassLoader classLoader) throws Exception {
+  private void ensureResourceReleaserIsCreatedByCorrectClassLoader(MuleArtifactClassLoader classLoader) throws Exception {
     assertThat(classLoader.getClass().getClassLoader(), is(currentThread().getContextClassLoader()));
     classLoader.setResourceReleaserClassLocation(TEST_RESOURCE_RELEASER_CLASS_LOCATION);
 
@@ -141,8 +168,8 @@ public class ResourceReleaserTestCase extends AbstractMuleTestCase {
 
   @Test
   public void cleanUpResourcesBundleFromDisposedClassLoader() throws Exception {
-    TestArtifactClassLoader classLoader =
-        new TestArtifactClassLoader(new TestArtifactClassLoader(currentThread().getContextClassLoader()));
+    TestPluginClassLoader classLoader =
+        new TestPluginClassLoader(classLoaderFactory.apply(currentThread().getContextClassLoader()));
     String resourceReleaserClassLocation = "/".concat(JdbcResourceReleaser.class.getName().replace(".", "/")).concat(".class");
     classLoader.setResourceReleaserClassLocation(resourceReleaserClassLocation);
 
@@ -166,9 +193,9 @@ public class ResourceReleaserTestCase extends AbstractMuleTestCase {
 
   @Test
   public void createsInstanceOnlyOnce() throws IOException {
-    try (TestArtifactClassLoader testArtifactClassLoader = new TestArtifactClassLoader(currentThread().getContextClassLoader())) {
-      ResourceReleaser firstInstance = testArtifactClassLoader.createResourceReleaserInstance();
-      ResourceReleaser secondInstance = testArtifactClassLoader.createResourceReleaserInstance();
+    try (MuleArtifactClassLoader testArtifactClassLoader = classLoaderFactory.apply(currentThread().getContextClassLoader())) {
+      ResourceReleaser firstInstance = ((KeepResourceReleaserInstance) testArtifactClassLoader).createResourceReleaserInstance();
+      ResourceReleaser secondInstance = ((KeepResourceReleaserInstance) testArtifactClassLoader).createResourceReleaserInstance();
 
       assertThat(firstInstance, sameInstance(secondInstance));
     }
@@ -177,13 +204,16 @@ public class ResourceReleaserTestCase extends AbstractMuleTestCase {
   private interface KeepResourceReleaserInstance {
 
     ResourceReleaser getResourceReleaserInstance();
+
+    ResourceReleaser createResourceReleaserInstance();
+
   }
 
-  private static class TestArtifactClassLoader extends MulePluginClassLoader implements KeepResourceReleaserInstance {
+  private static class TestPluginClassLoader extends MulePluginClassLoader implements KeepResourceReleaserInstance {
 
     private ResourceReleaser resourceReleaserInstance;
 
-    public TestArtifactClassLoader(ClassLoader parentCl) {
+    public TestPluginClassLoader(ClassLoader parentCl) {
       super("testId", new ArtifactDescriptor("test"), new URL[0], parentCl, new ClassLoaderLookupPolicy() {
 
         @Override
@@ -219,7 +249,59 @@ public class ResourceReleaserTestCase extends AbstractMuleTestCase {
     }
 
     @Override
-    protected ResourceReleaser createResourceReleaserInstance() {
+    public ResourceReleaser createResourceReleaserInstance() {
+      resourceReleaserInstance = super.createResourceReleaserInstance();
+      return resourceReleaserInstance;
+    }
+
+    @Override
+    public ResourceReleaser getResourceReleaserInstance() {
+      return resourceReleaserInstance;
+    }
+  }
+
+  private static class TestApplicationClassLoader extends MuleDeployableArtifactClassLoader
+      implements KeepResourceReleaserInstance {
+
+    private ResourceReleaser resourceReleaserInstance;
+
+    public TestApplicationClassLoader(ClassLoader parentCl) {
+      super("testId", new ArtifactDescriptor("test"), new URL[0], parentCl, new ClassLoaderLookupPolicy() {
+
+        @Override
+        public LookupStrategy getClassLookupStrategy(String className) {
+          return PARENT_FIRST;
+        }
+
+        @Override
+        public LookupStrategy getPackageLookupStrategy(String packageName) {
+          return null;
+        }
+
+        @Override
+        public ClassLoaderLookupPolicy extend(Map<String, LookupStrategy> lookupStrategies) {
+          return null;
+        }
+
+        @Override
+        public ClassLoaderLookupPolicy extend(Stream<String> packages, LookupStrategy lookupStrategy) {
+          return null;
+        }
+
+        @Override
+        public ClassLoaderLookupPolicy extend(Map<String, LookupStrategy> lookupStrategies, boolean overwrite) {
+          return null;
+        }
+
+        @Override
+        public ClassLoaderLookupPolicy extend(Stream<String> packages, LookupStrategy lookupStrategy, boolean overwrite) {
+          return null;
+        }
+      });
+    }
+
+    @Override
+    public ResourceReleaser createResourceReleaserInstance() {
       resourceReleaserInstance = super.createResourceReleaserInstance();
       return resourceReleaserInstance;
     }
